feat: refactor module system
