+++ app.js
@@ -1,4 +1,1 @@
-import './assets/javascript/server.js'
-import './assets/javascript/init.js';
 import './assets/javascript/modules/contentManager.js';
-import './assets/javascript/loadModules.js';
