--- assets/javascript/init.js
@@ -1,5 +1,0 @@
-const server = window.server;
-
-window.app = {};
-const app = window.app;
-app.modules = {};
