--- assets/javascript/loadModules.js
@@ -1,3 +1,0 @@
-Object.keys(app.modules).forEach((key) => {
-  app.modules[key].load();
-});
